requirements.txt
absl-py==2.4.0
accelerate==1.13.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.3.0
attrs==26.1.0
av==12.3.0
babel==2.18.0
backoff==2.2.1
beautifulsoup4==4.14.3
blake3==1.0.8
bleach==6.3.0
blinker==1.9.0
boto3==1.42.72
botocore==1.42.72
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
clip @ git+https://github.com/openai/CLIP.git@d05afc436d78f1c48dc0dbf8e5980a9d471f35f6
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
diffusers==0.37.0
docker==7.1.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.135.1
fastjsonschema==2.21.2
ffmpy==1.0.0
filelock==3.20.0
flatbuffers==25.12.19
fonttools==4.62.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.12.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.46
gradio==4.41.0
gradio_client==1.3.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.7.2
hydra-core==1.3.2
idna==3.11
ImageIO==2.37.3
importlib_metadata==9.0.0
importlib_resources==6.5.2
iopath==0.1.10
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.3
jmespath==1.1.0
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.2.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.5.0
lark==1.3.1
lazy-loader==0.5
lightning==2.6.0
lightning-utilities==0.15.3
lightning_sdk==2026.3.3
litai==0.0.10
litdata==0.2.58
litlogger==2026.3.9
litmodels==0.1.8
litserve==0.2.16
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mediapipe==0.10.33
mistune==3.2.0
mpmath==1.3.0
multidict==6.7.1
narwhals==2.18.1
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
obstore==0.9.2
omegaconf==2.3.0
opencv-contrib-python==4.13.0.92
opencv-python==4.10.0.84
orjson==3.11.7
packaging==23.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.6
peft==0.18.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.9.4
portalocker==3.2.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==4.25.9
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pycocotools==2.0.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.0.0
python-multipart==0.0.22
pytorch-lightning==2.6.0
pytz==2026.1.post1
PyYAML==6.0.1
pyzmq==27.1.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rpds-py==0.30.0
ruff==0.15.7
s3transfer==0.16.0
safetensors==0.7.0
sam2==1.1.0
scikit-image==0.24.0
scikit-learn==1.8.0
scipy==1.17.1
semantic-version==2.10.0
Send2Trash==2.1.0
setuptools==80.10.2
shellingham==1.5.4
simple-term-menu==1.6.6
six==1.17.0
smmap==5.0.3
sounddevice==0.5.5
soupsieve==2.8.3
stack-data==0.6.3
starlette==1.0.0
streamlit==1.32.0
streamlit-drawable-canvas==0.9.3
sympy==1.14.0
tabulate==0.10.0
tenacity==8.5.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.3.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.5.10
tinycss2==1.4.0
tokenizers==0.22.2
toml==0.10.2
tomlkit==0.12.0
torch==2.8.0
torchmetrics==1.9.0
torchvision==0.23.0+cu128
tornado==6.5.5
tqdm==4.66.4
traitlets==5.14.3
transformers==5.3.0
triton==3.4.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==12.0
Werkzeug==3.1.6
wheel==0.46.3
widgetsnbextension==4.0.15
yacs==0.1.8
yarl==1.23.0
zipp==3.23.0
